Add a basic GUI & various code clean-ups
Reverse the meaning of index 0/1 for the parameter that controls which
input to invert, so that the checkbox in the GUI, the native on/off
control that some hosts (*cough* Renoise) choose to display for a
two-option "choice" parameter instead of the actual choice labels, and
anywhere where it is converted to/from a bool all line up nicely.

Clean up audio parameter subclasses to use a common template base
class.

Add a basic GUI showing the two parameters, with linking back & forth
between the GUI and the actual underlying plugin parameter values.

Add a whole bunch of comments relating to future plans on how to maybe
implement instance linking.